requirements.txt
discord.py
google-api-python-client
psutil
aiohttp
feedparser
python-dotenv
